+++ lib/ui/map2/Map2HomePanel.dart
@@ -35,6 +35,7 @@ import 'package:illinois/service/Map2.dart';
 import 'package:illinois/service/Storage.dart';
 import 'package:illinois/service/StudentCourses.dart';
 import 'package:illinois/service/Wellness.dart';
+import 'package:illinois/ui/RootPanel.dart';
 import 'package:illinois/ui/events2/Event2HomePanel.dart';
 import 'package:illinois/ui/explore/ExploreMessagePopup.dart';
 import 'package:illinois/ui/map2/Map2BasePanel.dart';
@@ -52,6 +53,7 @@ import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/model/event2.dart';
 import 'package:rokwire_plugin/model/explore.dart';
 import 'package:rokwire_plugin/model/places.dart';
+import 'package:rokwire_plugin/service/Log.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
 import 'package:rokwire_plugin/service/connectivity.dart';
 import 'package:rokwire_plugin/service/events2.dart';
@@ -61,7 +63,6 @@ import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/places.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
-import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
 
 import '../widgets/HeaderBar.dart';
 
@@ -122,6 +123,7 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
   final GlobalKey _scaffoldKey = GlobalKey();
   final GlobalKey _contentHeadingBarKey = GlobalKey();
   final GlobalKey _contentTypesBarKey = GlobalKey();
+  final GlobalKey _contentTitleKey = GlobalKey();
   final GlobalKey _traySheetKey = GlobalKey();
   final GlobalKey _traySheetHeaderKey = GlobalKey();
   final GlobalKey _sortButtonKey = GlobalKey();
@@ -132,6 +134,7 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
   final GlobalKey _searchButtonKey = GlobalKey();
   final GlobalKey _paymentTypesButtonKey = GlobalKey();
   final GlobalKey _openNowButtonKey = GlobalKey();
+  final Map<Map2ContentType, GlobalKey> contentTypeKeys = {};
 
 
   final ScrollController _contentTypesScrollController = ScrollController();
@@ -167,6 +170,8 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
   Position? _currentLocation;
   Map<String, dynamic>? _mapStyles;
 
+  bool _needHeadingAccessibilityFocus = false;
+
   @override
   void initState() {
     NotificationService().subscribe(this, [
@@ -178,7 +183,8 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
       Map2ExplorePOICard.notifyPOIUpdated,
       Map2.notifySelect,
       FlexUI.notifyChanged,
-      uiuc.TabBar.notifySelectionChanged,
+      RootPanel.notifyTabAppear,
+      RootPanel.notifyTabDisappear,
     ]);
 
     _availableContentTypes = Map2ContentTypeImpl.availableTypes;
@@ -192,6 +198,7 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
     _initExplores();
     WidgetsBinding.instance.addPostFrameCallback((_) {
       _accessibilityFocusHeading();
+      _needHeadingAccessibilityFocus = true;
     });
     super.initState();
   }
@@ -251,10 +258,15 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
       _updateAvailableContentTypes();
       updateLocationServicesStatus();
     }
-    else if (name == uiuc.TabBar.notifySelectionChanged){
-      int index = JsonUtils.intValue(param) ?? 0;
-      if(index == 1)  //Index to code ? or pass directly code
+    else if (name == RootPanel.notifyTabAppear){
+      if ((JsonUtils.cast<RootTab>(param) == RootTab.Map) && mounted) {
         _accessibilityFocusHeading();
+      }
+    }
+    else if (name == RootPanel.notifyTabDisappear){
+      if ((JsonUtils.cast<RootTab>(param) == RootTab.Map) && mounted) {
+        _clearContent();
+      }
     }
   }
 
@@ -527,6 +539,16 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
     }
   }
 
+  @override
+  onMapCreated(GoogleMapController controller){
+    super.onMapCreated(controller);
+    if(_needHeadingAccessibilityFocus){
+      Log.d("_accessibilityFocusHeading");
+      _accessibilityFocusHeading();
+      _needHeadingAccessibilityFocus = false;
+    }
+  }
+
   // Content Types
 
   Widget get _contentTypesBar => 
@@ -546,7 +568,7 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
       if (_availableContentTypes.contains(contentType)) {
         entries.add(Padding(
           padding: EdgeInsets.only(left: entries.isNotEmpty ? 8 : 0),
-          child: Map2ContentTypeButton(contentType.displayTitle,
+          child: Map2ContentTypeButton(key: contentTypeKeys[contentType] ??= GlobalKey(), contentType.displayTitle,
             onTap: () => _onTapContentTypeEntry(contentType),
           )
         ));
@@ -577,7 +599,7 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
 
   void _onContentTypeEntry(Map2ContentType contentType) {
     setState(() {
-      Storage().storedMap2ContentType = _selectedContentType = contentType;
+      _selectedContentType = contentType;
     });
     _initExplores();
   }
@@ -626,18 +648,20 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
   // Content Filters
 
   Widget get _contentHeadingBar =>
-    Container(key: _contentHeadingBarKey, decoration: _contentHeadingDecoration, child:
-      Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: _searchOn ? <Widget>[
-        _contentFilterSearchBar,
-      ] : <Widget>[
-        _contentTitleBar,
-        if ((_exploresProgress == null) || (_exploresProgress == ExploreProgressType.update))
-          ...[_contentFilterButtonsBar ?? Container(),
-            ...(_contentFilterButtonsExtraBars ?? [])
-          ],
-        if (_exploresProgress == null)
-          _contentFilterDescriptionBar ?? Container(),
-      ],),
+    Semantics(key: _contentHeadingBarKey, child:
+      Container(decoration: _contentHeadingDecoration, child:
+        Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: _searchOn ? <Widget>[
+          _contentFilterSearchBar,
+        ] : <Widget>[
+          _contentTitleBar,
+          if ((_exploresProgress == null) || (_exploresProgress == ExploreProgressType.update))
+            ...[_contentFilterButtonsBar ?? Container(),
+              ...(_contentFilterButtonsExtraBars ?? [])
+            ],
+          if (_exploresProgress == null)
+            _contentFilterDescriptionBar ?? Container(),
+        ],),
+      )
     );
 
   BoxDecoration get _contentHeadingDecoration =>
@@ -648,7 +672,7 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
     );
 
   Widget get _contentTitleBar =>
-    Semantics(header: true, container: true, child:
+    Semantics(key: _contentTitleKey, header: true, container: true, focused: true, child:
       Row(children: [
         Expanded(child:
           Padding(padding: EdgeInsets.only(left: 16, top: 8, bottom: 8), child:
@@ -668,31 +692,35 @@ class _Map2HomePanelState extends Map2BasePanelState<Map2HomePanel>
 
   void _onTapClearContentType() {
     Analytics().logSelect(target: 'Content: Clear');
-    setState(() {
-      Storage().storedMap2ContentType = _selectedContentType = null;
-      _explores = _filteredExplores = null;
-      _selectedExploreGroup = null;
-      _trayExplores = null;
-      _exploresTask = null;
-      _exploresProgress = null;
-
-      _storiedSitesTags = null;
-      _expandedStoriedSitesTag = null;
-
-      markers = null;
-      exploreMapGroups = null;
-      targetCameraUpdate = null;
-      buildMarkersTask = null;
-      lastMapZoom = null;
-      markersProgress = false;
-
-      _pinnedExplore = null;
-      _pinnedMarker = null;
-    });
+    GlobalKey? _lastSelectedContentKey = contentTypeKeys[_selectedContentType];
+    _clearContent();
     WidgetsBinding.instance.addPostFrameCallback((_){
       _updateContentTypesScrollPosition();
+      AppSemantics.triggerAccessibilityFocus(_lastSelectedContentKey);
     });
   }
+
+  void _clearContent() => setState((){
+    _selectedContentType = null;
+    _explores = _filteredExplores = null;
+    _selectedExploreGroup = null;
+    _trayExplores = null;
+    _exploresTask = null;
+    _exploresProgress = null;
+
+    _storiedSitesTags = null;
+    _expandedStoriedSitesTag = null;
+
+    markers = null;
+    exploreMapGroups = null;
+    targetCameraUpdate = null;
+    buildMarkersTask = null;
+    lastMapZoom = null;
+    markersProgress = false;
+
+    _pinnedExplore = null;
+    _pinnedMarker = null;
+  });
 
   // My Locactions Content && Selection
 
@@ -1186,7 +1214,7 @@ extension _Map2HomePanelFilters on _Map2HomePanelState {
                 onTap: _onTapShareFilter
               )
             )),
-            IndexedSemantics(index: 1, child: Semantics( container: true, child:
+            IndexedSemantics(index: 3, child: Semantics( container: true, child:
               Map2PlainImageButton(imageKey: 'close',
                   label: Localization().getStringEx('panel.events2.home.bar.button.clear.title', 'Clear Filters'),
                   hint: Localization().getStringEx('panel.events2.home.bar.button.clear.hint', 'Tap to clear current filters'),
@@ -1883,9 +1911,7 @@ extension _Map2HomePanelFilters on _Map2HomePanelState {
         _selectedSortOrder = value.right ?? _expectedSortOrder(value.left);
       });
       _onSortChanged();
-      Future.delayed(Duration(seconds: Platform.isIOS ? 1 : 0), () =>
-        AppSemantics.triggerAccessibilityFocus(_sortButtonKey)
-      );
+      AppSemantics.triggerAccessibilityFocus(_sortButtonKey, delay: Duration(seconds: 1));
     }
   }
 
@@ -2122,6 +2148,6 @@ extension _Map2Accessibility on _Map2HomePanelState{
   void _accessibilityFocusHeading() {
     AppSemantics.triggerAccessibilityFocus(_rootHeaderBarTitleKey); //When already on this tab
     WidgetsBinding.instance.addPostFrameCallback((_) => //When coming from other tab
-      AppSemantics.triggerAccessibilityFocus(_rootHeaderBarTitleKey));
+      AppSemantics.triggerAccessibilityFocus(_rootHeaderBarTitleKey, delay: Duration(microseconds: 200))); // Delay so we can determine properly if it's already focused so we can avoid second pronunciation
   }
 }