+++ lib/ui/widgets/TabBar.dart
@@ -79,16 +79,6 @@ class TabBar extends rokwire.TabBar {
         onTap: (rokwire.TabWidget tabWidget) => _onSwitchTab(index, tabWidget),
       ));
     }
-    else if (code == 'map0') {
-      return FocusableSemanticsWidget(onSelect: (rokwire.TabWidget tabWidget) => _onSwitchTab(index, tabWidget), child: rokwire.TabWidget(
-        label: Localization().getStringEx('tabbar.map0.title', 'Map0'),
-        hint: Localization().getStringEx('tabbar.map0.hint', 'Map0 Page'),
-        iconKey: 'location-outline-gray',
-        selectedIconKey: 'location',
-        selected: (tabController?.index == index),
-        onTap: (rokwire.TabWidget tabWidget) => _onSwitchTab(index, tabWidget),
-      ));
-    }
     else if (code == 'academics') {
       return FocusableSemanticsWidget(onSelect: (rokwire.TabWidget tabWidget) => _onSwitchTab(index, tabWidget), child: rokwire.TabWidget(
         label: Localization().getStringEx('tabbar.academics.title', 'Academics'),