+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Fixed
+- Fixed navigation to Assistant, Academics and Wellness home panels [#5541](https://github.com/rokwire/illinois-app/issues/5541).
+- Do not show "View In-Person Courses on Map" for online courses [#5583](https://github.com/rokwire/illinois-app/issues/5583).
+- WEB: Accessibility for the mobile devices panel [#5594](https://github.com/rokwire/illinois-app/issues/5594).
+
+## [7.3.34] - 2025-12-12
+### Fixed
 - ILLordle: Fixed daily word processing [#5442](https://github.com/rokwire/illinois-app/issues/5442).
 - Assistant: increase UI fragments width and fix paddings [#5542](https://github.com/rokwire/illinois-app/issues/5542).
 - WEB: ILLordle: Fixed loading data [#5568](https://github.com/rokwire/illinois-app/issues/5568).
@@ -14,8 +20,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sort dining schedules by time [#5577](https://github.com/rokwire/illinois-app/issues/5577).
 - WEB: Show videos in a Group post [#5513](https://github.com/rokwire/illinois-app/issues/5513).
 - Accessibility for unlabeled buttons [#5573](https://github.com/rokwire/illinois-app/issues/5573).
+- ProfileInfoEditPanel Improved accessibility for buttons with same label [#4874](https://github.com/rokwire/illinois-app/issues/4874).
 - Image Alt text is required [#4836](https://github.com/rokwire/illinois-app/issues/4836).
 - Image Alt text for GameDetails [#5553](https://github.com/rokwire/illinois-app/issues/5553).
+- Improve accessibility focus for map panel [#5500](https://github.com/rokwire/illinois-app/issues/5500).
+### Changed
+- Removed default map selection to Campus Buildings [#5580](https://github.com/rokwire/illinois-app/issues/5580).
+- Show Map2 without last selected content [#5580](https://github.com/rokwire/illinois-app/issues/5580).
+- Load default camera position from app config [#5580](https://github.com/rokwire/illinois-app/issues/5580).
+- Improved usability of Student Courses detail panel [#5583](https://github.com/rokwire/illinois-app/issues/5583).
+### Fixed
+- Improved speech to text management in Assistant panel for iOS [#4364](https://github.com/rokwire/illinois-app/issues/4364).
+### Added
+- WEB: Do not allow web app usage for mobile devices [#5594](https://github.com/rokwire/illinois-app/issues/5594).
 
 ## [7.3.33] - 2025-12-01
 ### Added